Lists of terms tab in a different component
